NDS-228: moves the example global CSS code to the sandbox
